cleaning up compute class methods
